increment java version to 17 in ci/cd
