+++ .github/workflows/android_ci.yml
@@ -16,7 +16,7 @@ jobs:
       - name: Set up JDK environment(JDK 11)
         uses: actions/setup-java@v1.4.3
         with:
-          java-version: 11
+          java-version: 17
 
       - name: Generate Release APK
         run: ./gradlew assembleRelease
