+++ data/src/main/java/io/github/turskyi/data/datasources/local/entities/CountryEntity.kt
@@ -8,10 +8,10 @@ data class CountryEntity(
         val shortName: String,
         val name: String,
         val flag: String,
-        // parameter to use on the "all countries" page, to show if the country was already visited
+        /** parameter to use on the "all countries" page, to show if the country was already visited*/
         @field:JvmField
         val isVisited: Boolean,
-){
+) {
         // required empty constructor for firestore serialization
         constructor() : this(0, "", "", "", false)
 }
