+++ presentation/src/main/AndroidManifest.xml
@@ -21,7 +21,9 @@
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_WIFI_STATE" />
-    <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
+    <uses-permission
+        android:name="android.permission.READ_EXTERNAL_STORAGE"
+        android:maxSdkVersion="32" />
     <uses-permission
         android:name="com.google.android.gms.permission.AD_ID"
         tools:node="remove" />
@@ -56,7 +58,7 @@
         </activity>
 
         <activity
-            android:name=".features.allcountries.view.ui.AllCountriesActivity"
+            android:name=".features.countries.view.ui.AllCountriesActivity"
             android:parentActivityName=".features.home.view.ui.HomeActivity"
             android:theme="@style/AppTheme.NoActionBar"
             android:windowSoftInputMode="stateAlwaysHidden" />
