+++ presentation/src/main/java/io/github/turskyi/travellingpro/features/home/view/ui/HomeActivity.kt
@@ -11,7 +11,11 @@ import android.os.Handler
 import android.os.Looper
 import android.os.SystemClock
 import android.provider.Settings.ACTION_WIRELESS_SETTINGS
-import android.view.*
+import android.view.Gravity
+import android.view.Menu
+import android.view.MenuInflater
+import android.view.MenuItem
+import android.view.View
 import android.window.OnBackInvokedDispatcher
 import androidx.activity.result.ActivityResult
 import androidx.activity.result.ActivityResultLauncher
@@ -33,7 +37,7 @@ import io.github.turskyi.travellingpro.entities.City
 import io.github.turskyi.travellingpro.entities.Country
 import io.github.turskyi.travellingpro.entities.VisitedCountry
 import io.github.turskyi.travellingpro.entities.VisitedCountryNode
-import io.github.turskyi.travellingpro.features.allcountries.view.ui.AllCountriesActivity
+import io.github.turskyi.travellingpro.features.countries.view.ui.AllCountriesActivity
 import io.github.turskyi.travellingpro.features.flags.view.FlagsActivity
 import io.github.turskyi.travellingpro.features.flags.view.FlagsActivity.Companion.EXTRA_ITEM_COUNT
 import io.github.turskyi.travellingpro.features.flags.view.FlagsActivity.Companion.EXTRA_POSITION
@@ -43,7 +47,17 @@ import io.github.turskyi.travellingpro.features.home.viewmodels.HomeActivityView
 import io.github.turskyi.travellingpro.features.travellers.view.TravellersActivity
 import io.github.turskyi.travellingpro.utils.Event
 import io.github.turskyi.travellingpro.utils.decoration.SectionAverageGapItemDecoration
-import io.github.turskyi.travellingpro.utils.extensions.*
+import io.github.turskyi.travellingpro.utils.extensions.action
+import io.github.turskyi.travellingpro.utils.extensions.getScreenWidth
+import io.github.turskyi.travellingpro.utils.extensions.isOnline
+import io.github.turskyi.travellingpro.utils.extensions.mapVisitedCountryNodeToCountry
+import io.github.turskyi.travellingpro.utils.extensions.openActivity
+import io.github.turskyi.travellingpro.utils.extensions.openActivityWithArgs
+import io.github.turskyi.travellingpro.utils.extensions.openInfoDialog
+import io.github.turskyi.travellingpro.utils.extensions.showSnackBar
+import io.github.turskyi.travellingpro.utils.extensions.showSnackWithAction
+import io.github.turskyi.travellingpro.utils.extensions.toast
+import io.github.turskyi.travellingpro.utils.extensions.toastLong
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.launch
